INNER: the list-zip-threshold is config item

+++ conf/config.go
@@ -17,7 +17,7 @@ type Server struct {
 	Auth             string `cfg:"auth;;;client connetion auth"`
 	Listen           string `cfg:"listen; 0.0.0.0:7369; netaddr; address to listen"`
 	MaxConnection    int64  `cfg:"max-connection;1000;numeric;client connection count"`
-	ListZipThreshold int64  `cfg:"list-zip-threshold;100;numeric;the max limit length of elements in list"`
+	ListZipThreshold int    `cfg:"list-zip-threshold;100;numeric;the max limit length of elements in list"`
 }
 
 //Tikv config is the config of tikv sdk
